feature: added menu; refactor: moved classes to separate files
